Add state saving function

+++ qclobot/qcframe.py
@@ -23,7 +23,6 @@ import os
 import logging
 import math
 from collections import OrderedDict
-import shelve
 import shutil
 
 try:
@@ -76,21 +75,54 @@ class QcFrame(object):
 
     def __del__(self):
         self._save()
-        pass
-        
+
+    # save & load ------------------------------------------------------
     def _load(self):
-        return
-        data = shelve.open(os.path.join(self.work_dir, 'qcframe'))
-        self._fragments = data.get('fragments', {})
-        data.close()
+        path = os.path.join(self.work_dir, 'qcframe.mpac')
+        if os.path.exists(path):
+            self._logger.info('load the fragment state: {}'.format(path))
+            f = open(path, 'rb')
+            packed = f.read()
+            state_dat = msgpack.unpackb(packed)
+            f.close()
+            self.set_by_raw_data(state_dat)
+        else:
+            self._logger.info('not found the state file')
         
     def _save(self):
-        return 
-        data = shelve.open(os.path.join(self.work_dir, 'qcframe'))
-        for k, v in self._fragments.items():
-            print(k, v)
-        data['fragments'] = self._fragments
-        data.close()
+        path = os.path.join(self.work_dir, 'qcframe.mpac')
+        self._logger.info('save the fragment state: {}'.format(path))
+
+        state_dat = self.get_raw_data()
+        packed = msgpack.packb(state_dat)
+        f = open(path, 'wb')
+        f.write(packed)
+        f.close()
+
+    def get_raw_data(self):
+        return self.__getstate__()
+        
+    def set_by_raw_data(self, raw_data):
+        self.__setstate__(raw_data)
+        
+    def __getstate__(self):
+        state = {}
+        state['name'] = self.name
+        tmp_frgs = []
+        for k, frg in self.fragments():
+            tmp_frgs.append((k, frg.get_raw_data()))
+        state['fragments'] = tmp_frgs
+        state['state'] = self._state
+        return state
+
+    def __setstate__(self, state):
+        assert(isinstance(state, dict))
+        self._name = state.get('name')
+        self._fragments = OrderedDict()
+        if 'fragments' in state:
+            for (k, frg) in state.get('fragments'):
+                self._fragments[k] = qclo.QcFragment(frg, qc_parent=self)
+        self._state = state.get('state', {})
         
     # pdfparam ---------------------------------------------------------
     def _get_pdfparam(self):
@@ -110,7 +142,6 @@ class QcFrame(object):
             else:
                 pdfsim = pdf.PdfSim()
                 self._pdfparam = pdf.get_default_pdfparam()
-                #self._pdfparam.save(pdfparam_path)
         return self._pdfparam
 
     pdfparam = property(_get_pdfparam)
@@ -176,7 +207,7 @@ class QcFrame(object):
         assert(len(self.name) > 0)
         self._work_dir = os.path.abspath(os.path.join(os.curdir, self.name))    
         if not os.path.exists(self.work_dir):
-            self._logger.info('make workdir: {}'.format(self.work_dir))
+            self._logger.info('make work dir: {}'.format(self.work_dir))
             os.mkdir(self.work_dir)
         else:
             self._logger.debug('already exist: {}'.format(self.work_dir))
@@ -188,6 +219,8 @@ class QcFrame(object):
         self._logger.info('=' * 20)
         self._logger.info('>>>> {job_name}@{frame_name}'.format(job_name = job_name,
                                                                 frame_name = self.name))
+        self._logger.info('work dir: {work_dir}'.format(work_dir=self.work_dir))
+        
         self._logger.info('=' * 20)
         self._prev_dir = os.path.abspath(os.curdir)
         os.chdir(self.work_dir)
@@ -216,6 +249,28 @@ class QcFrame(object):
     # ==================================================================
     # STATE
     # ==================================================================
+    # guess_density ----------------------------------------------------
+    def _get_state_finished_guess_density(self):
+        self._state.setdefault('is_finished_guess_density', False)
+        return self._state['is_finished_guess_density']
+
+    def _set_state_finished_guess_density(self, yn):
+        self._state['is_finished_guess_density'] = bool(yn)
+        
+    is_finished_guess_density = property(_get_state_finished_guess_density,
+                                         _set_state_finished_guess_density)
+        
+    # guess_QCLO -------------------------------------------------------
+    def _get_state_finished_guess_QCLO(self):
+        self._state.setdefault('is_finished_guess_QCLO', False)
+        return self._state['is_finished_guess_QCLO']
+
+    def _set_state_finished_guess_QCLO(self, yn):
+        self._state['is_finished_guess_QCLO'] = bool(yn)
+        
+    is_finished_guess_QCLO = property(_get_state_finished_guess_QCLO,
+                                      _set_state_finished_guess_QCLO)
+        
     # pre-SCF ----------------------------------------------------------
     def _get_state_finished_prescf(self):
         self._state.setdefault('is_finished_prescf', False)
@@ -237,19 +292,56 @@ class QcFrame(object):
         
     is_finished_scf = property(_get_state_finished_scf,
                                _set_state_finished_scf)
+
+    # pick density matrix  ---------------------------------------------
+    def _get_state_finished_pickup_density_matrix(self):
+        self._state.setdefault('is_finished_pickup_density_matrix', False)
+        return self._state['is_finished_pickup_density_matrix']
+
+    def _set_state_finished_pickup_density_matrix(self, yn):
+        self._state['is_finished_pickup_density_matrix'] = bool(yn)
         
+    is_finished_pickup_density_matrix = property(_get_state_finished_pickup_density_matrix,
+                                                 _set_state_finished_pickup_density_matrix)
+    
+    # LO ---------------------------------------------------------------
+    def _get_state_finished_LO(self):
+        self._state.setdefault('is_finished_LO', False)
+        return self._state['is_finished_LO']
+
+    def _set_state_finished_LO(self, yn):
+        self._state['is_finished_LO'] = bool(yn)
+        
+    is_finished_LO = property(_get_state_finished_LO,
+                               _set_state_finished_LO)
+    
+    # pickup LO --------------------------------------------------------
+    def _get_state_finished_pickup_LO(self):
+        self._state.setdefault('is_finished_pickup_LO', False)
+        return self._state['is_finished_pickup_LO']
+
+    def _set_state_finished_pickup_LO(self, yn):
+        self._state['is_finished_pickup_LO'] = bool(yn)
+        
+    is_finished_pickup_LO = property(_get_state_finished_pickup_LO,
+                                     _set_state_finished_pickup_LO)
     # ==================================================================
     # GUESS
     # ==================================================================
     # guess density ----------------------------------------------------
     def guess_density(self, run_type ='rks'):
+        if self.is_finished_guess_density:
+            self._logger.info('guess_density has been calced.')
+            return
+
         self.cd_work_dir('guess_density')
+
         guess_density_matrix_path = 'guess.density.{}.mat'.format(run_type)
 
         # 既存のデータを消去する
         if os.path.exists(guess_density_matrix_path):
             os.remove(guess_density_matrix_path)
-        
+                
         pdfsim = pdf.PdfSim()
         pdfsim.setup()
         
@@ -280,10 +372,17 @@ class QcFrame(object):
         # check
         self._check_path(guess_density_matrix_path)
         
+        self.is_finished_guess_density = True
+        self._save()
+            
         self.restore_cwd()
 
         
     def guess_QCLO(self, run_type='rks', isCalcOrthogonalize = False):
+        if self.is_finished_guess_QCLO:
+            self._logger.info('guess_density has been calced.')
+            return
+
         self.cd_work_dir('guess_QCLO')
 
         guess_QCLO_matrix_path = 'guess.QCLO.{}.mat'.format(run_type)
@@ -322,11 +421,14 @@ class QcFrame(object):
         # check
         self._check_path(guess_QCLO_matrix_path)
 
-        self.create_occupation_file(run_type)
+        self._create_occupation_file(run_type)
         
+        self.is_finished_guess_QCLO = True
+        self._save()
+
         self.restore_cwd()
 
-    def create_occupation_file(self, run_type='rks'):
+    def _create_occupation_file(self, run_type='rks'):
         self.cd_work_dir('create occ')
 
         occ_level = -1
@@ -346,7 +448,8 @@ class QcFrame(object):
         occ_vtr_path = 'guess.occ.{}.vtr'.format(run_type.lower())
         occ_vtr.save(occ_vtr_path)
         self._check_path(occ_vtr_path)
-            
+
+        self._save()
         self.restore_cwd()
         
     # ==================================================================
@@ -356,10 +459,13 @@ class QcFrame(object):
     def calc_preSCF(self, dry_run=False):
         '''
         '''
-        self.cd_work_dir('calc SP')
+        if self.is_finished_prescf:
+            self._logger.info('preSCF has been calced.')
+            return
+        
+        self.cd_work_dir('calc preSCF')
 
         pdfsim = pdf.PdfSim()
-
         for frg_name, frg in self.fragments():
             frg.set_basisset(self.pdfparam)
         self.pdfparam.molecule = self.frame_molecule
@@ -369,42 +475,45 @@ class QcFrame(object):
                   workdir = self.work_dir,
                   db_path = self.db_path,
                   dry_run = dry_run)
-
+        
         self._pdfparam = None # cache clear
         self.is_finished_prescf = True
-        
+        self._save()
+
         self.restore_cwd()
-    
+        
     # sp ---------------------------------------------------------------
     def calc_sp(self, dry_run=False):
         '''
         calculate single point energy
         '''
+        if self.is_finished_scf:
+            self._logger.info('SP has been calced.')
+            return
+            
+        if self.is_finished_prescf != True:
+            self.calc_preSCF(dry_run)
+
         self.cd_work_dir('calc SP')
 
         pdfsim = pdf.PdfSim()
-
         for frg_name, frg in self.fragments():
             frg.set_basisset(self.pdfparam)
         self.pdfparam.molecule = self.frame_molecule
 
         #self.output_xyz("{}/model.xyz".format(self.name))
 
-        step_control = 'guess scf'
-        if self.is_finished_prescf != True:
-            step_control = 'integral ' + step_control
-        self.pdfparam.step_control = step_control
-        
+        self.pdfparam.step_control = 'guess scf'
         pdfsim.sp(self.pdfparam,
                   workdir = self.work_dir,
                   db_path = self.db_path,
                   dry_run = dry_run)
-
+            
         self._pdfparam = None # cache clear
-        self.is_finished_prescf = True
         self.is_finished_scf = True
         self._switch_fragments()
-        
+        self._save()
+            
         self.restore_cwd()
 
     # ------------------------------------------------------------------
@@ -416,10 +525,14 @@ class QcFrame(object):
         '''
         密度行列を各フラグメントに割り当てる
         '''
+        if self.is_finished_pickup_density_matrix:
+            self._logger.info('pickup density matrix has done.')
+            return
+        
         self.cd_work_dir('pickup density matrix')
         
         dens_mat_path = self.pdfparam.get_density_matrix_path(runtype=runtype)
-        self._logger.info('dens.mat={}'.format(dens_mat_path))
+        self._logger.info('ref. density matrix: {}'.format(dens_mat_path))
         
         global_dim = 0
         for frg_name, frg in self.fragments():
@@ -427,15 +540,14 @@ class QcFrame(object):
             if dim > 0:
                 frg_dens_mat_path = 'Ppq.{}.{}.mat'.format(runtype, frg_name)
                 self._logger.info(
-                    'density matrix select frg:{} in frame:{} was saved as {} ({} -> {})'.format(
-                        frg_name,
-                        self.name,
-                        frg_dens_mat_path,
-                        global_dim,
-                        global_dim +dim -1))
+                    'select [{start}:{end}] for [{frame}][{fragment}]'.format(
+                        frame=frg_name,
+                        fragment=self.name,
+                        start=global_dim,
+                        end=global_dim +dim -1))
 
                 # フラグメント対応部分を切り出す
-                pdf.run_pdf(['mat-select', '-v',
+                pdf.run_pdf(['mat-select',
                              '-t', global_dim,
                              '-l', global_dim,
                              '-b', global_dim +dim -1,
@@ -452,25 +564,37 @@ class QcFrame(object):
                 frg.set_density_matrix(frg_dens_mat_path)
 
                 self._logger.info(
-                    'density matrix of the frg:{} in frame:{} was saved as {}'.format(
-                        frg_name,
-                        self.name,
-                        frg_dens_mat_path))
+                    'density matrix for [{frame}][{fragment}] was saved as {path}'.format(
+                        fragment=frg_name,
+                        frame=self.name,
+                        path=frg_dens_mat_path))
                 global_dim += dim
 
+        self._logger.is_finished_pickup_density_matrix = True
+        self._save()
         self.restore_cwd()
             
     # ------------------------------------------------------------------
     def calc_lo(self):
+        if self.is_finished_LO:
+            self._logger.info('LO has done.')
+            return
+
         self.cd_work_dir('calc lo')
 
         self._logger.info('start lo calculation.')
         pdf.run_pdf('lo')
 
+        self.is_finished_LO = True
+        self._save()
         self.restore_cwd()
 
     # ------------------------------------------------------------------
     def pickup_lo(self):
+        if self.is_finished_pickup_LO:
+            self._logger.info('pickup LO has been finished.')
+            return
+
         self.cd_work_dir('pickup lo')
         
         # debug
@@ -552,6 +676,8 @@ class QcFrame(object):
         self._trans_LO()
 
         # finish
+        self.is_finished_pickup_LO = True
+        self._save()
         self.restore_cwd()
 
     def _get_AO_fragment_table(self, num_of_AOs):
